moved plane classes into package Planes and updated affected classes
