+++ src/main/java/com/example/demo/Levels/LevelOne.java
@@ -1,7 +1,7 @@
 package com.example.demo.Levels;
 
 import com.example.demo.Actors.ActiveActorDestructible;
-import com.example.demo.Actors.EnemyPlane;
+import com.example.demo.Planes.EnemyPlane;
 import com.example.demo.LevelViews.LevelView;
 
 public class LevelOne extends LevelParent {
