+++ src/main/java/com/example/demo/Levels/LevelParent.java
@@ -1,9 +1,9 @@
 package com.example.demo.Levels;
 
 import com.example.demo.Actors.ActiveActorDestructible;
-import com.example.demo.Actors.FighterPlane;
+import com.example.demo.Planes.FighterPlane;
 import com.example.demo.LevelViews.LevelView;
-import com.example.demo.Actors.UserPlane;
+import com.example.demo.Planes.UserPlane;
 import javafx.animation.KeyFrame;
 import javafx.animation.Timeline;
 import javafx.scene.Group;
